Rewrite VersionHistory to make use of ModalV2 instead of Lightbox
